==> Price Insight System/1.Airbnb_Raw-Listing_Cleansing.ipynb <==
## Group

#### Hi! My group consists of me, Lincoln and Leong Seng. This is a write-up of our data science project for SC1015 Introduction to Data Science & Artificial Intelligence. 

#### Nowadays, data are being produced in an enormous amount. With the vast data available, these data can be collected, allowing us to make meaningful decisions by infering from the results of our analysis. With this vast amount of data and it's constant increase, visualization tools enable us to deliver our message to our audience and inform them about our findings.

#### Several varaiable have missing/nan variable, they are replace with certain methods
__new_host_acceptance_rate is replace to 0 if and only if 'number_of_reviews' are 0__
__new_host_acceptance_rate NaN value are replace with mean with review is not 0__


#### We will be exploring and visualizing the dataset from Airbnb in Singapore using exploratory data analysis techiques. Follow my group as we find out the distribution of every Airbnb listing based on the price range, room type, review, and other related factors.

### What is Exploratory Data Analysis?

#### Exploratory data analysis (EDA), also known as Data Exploration is part of the Data Analysis process whereby several techniques are used to better understand the dataset being used.

#### Some of this techniques are:
* Extrating important variables and ignoring useless variables
* Identifying outliers, filling in missing values
* Understanding our data, maximizing our insights on a dataset and minimizing potential error that may occur later in the process

#### Main components of Exploratory Data Analysis:
* Accquire and loading data
* Cleaning the dataset
* Exploring and Visualizing Data

### Load Python libraries

# Basic Libraries
import numpy as np
import pandas as pd
import seaborn as sb
import matplotlib.pyplot as plt # we only need pyplot
sb.set() # set the default Seaborn style for graphics

### Load dataset

# Import the dataset

Listing_RawData = pd.read_csv("listing_raw.csv")
Listing_df = pd.DataFrame(Listing_RawData)

Listing_RawData.head(5)

### Listing the dtype of each column

# For column reference: https://docs.google.com/spreadsheets/d/1iWCNJcSutYqpULSQHlNyGInUvHg2BoUGoNRIGa6Szc4/edit#gid=982310896

Listing_RawData.info()

Let's list down all our observations from the super-quick glance of the dataset, as above.

* There are `73` variables/features/columns and `3672` observations/samples/rows in the dataset.

* The response variable is `price`

Possible Predictors to consider:
* `host_acceptance_rate` & `host_is_superhost`
* `neighbourhood_cleansed` & `neighbourhood_group_cleansed`
* `latitude` & `longitude`
*  `room_type`
* `bedrooms` & `beds`
* `number_of_reviews`
* `availability_365`

Predictor definition: https://docs.google.com/spreadsheets/d/1iWCNJcSutYqpULSQHlNyGInUvHg2BoUGoNRIGa6Szc4/edit#gid=982310896

### Displaying the columns `host_acceptance_rate` & `number_of_reviews`

Listing_df[['host_acceptance_rate', 'number_of_reviews']].describe

#replace host_acceptance_rate NaN values with 0 for records with 0 reviews :D

Listing_df['new_host_acceptance_rate'] = Listing_df['host_acceptance_rate']
Listing_df['check_for_nan'] = Listing_df['host_acceptance_rate'].isnull()
Listing_df['check_for_nan'] = Listing_df['check_for_nan'].replace({True: 111, False: 222})



new_df = Listing_df.loc[Listing_df['check_for_nan']==111]
new_df =  new_df.loc[Listing_df['number_of_reviews'] == 0]
new_df['new_host_acceptance_rate'] = new_df['new_host_acceptance_rate'].fillna(0)


##update_dataframe
Listing_df.update(new_df)         

#### Setting `new_host_acceptance_rate` to 0 if and only if `number_of_reviews` are 0

Listing_df[['new_host_acceptance_rate', 'number_of_reviews']].describe

#Remove host_acceptance_rate %
Listing_df['new_host_acceptance_rate'].replace("%", "", inplace=True, regex=True)

#Convert new_host_acceptance_rate dtype from object to float
#describe new_host_acceptance_rate

Listing_df['new_host_acceptance_rate'] = pd.to_numeric(Listing_df['new_host_acceptance_rate'])

Listing_df['new_host_acceptance_rate'].describe()

#obtain mean from new_host_acceptance_rate
median_host_acceptance_rate = Listing_df['new_host_acceptance_rate'].median()

#replace new_host_acceptance_rate NaN with mean
Listing_df['new_host_acceptance_rate'].replace(np.nan, median_host_acceptance_rate, inplace=True, regex=True)

#after replacing NaN with mean
Listing_df['new_host_acceptance_rate'].describe()

### Replacing NULL with relevant values. Removing unrequired characters.

#Replace reviews_per_month NaN with 0
Listing_df['reviews_per_month'].replace(np.nan, 0, inplace=True, regex=True)

#replace host_is_superhost NaN with f
Listing_df['host_is_superhost'].replace(np.nan, 'f', inplace=True, regex=True)

#Replace bedroom NaN with 0
Listing_df['bedrooms'].replace(np.nan, 0, inplace=True, regex=True)

#Replace beds NaN with 0
Listing_df['beds'].replace(np.nan, 0, inplace=True, regex=True)

#Remove price $ 
Listing_df['price'].replace("\$|\$", "", inplace=True, regex=True)

Listing_df['price'].replace(",", "", inplace=True, regex=True)

#Convert price dtype from object to float
#describe price

Listing_df['price'] = pd.to_numeric(Listing_df['price'])

#Row 983 to be dropped - pricing is $10,286.00 [outlier]
#Row 743 to be dropped - pricing is $8900.00 [outlier] (NEED TO CHECK)
Listing_df = Listing_df[Listing_df.price < 10000]

Listing_df['price'].describe()

Listing_df[['new_host_acceptance_rate', 'host_is_superhost', 'bedrooms', 'beds', 'price']]

### Exporting the dataframe into a new csv

#Export columns to new csv for easier analization 

# 1) Check drive location
import os
cwd = os.getcwd()
saved_file = cwd + "\PredictPrice.csv"

# 2) New dataframe for export
export_df = Listing_df[['id', 'new_host_acceptance_rate', 'host_is_superhost', 'neighbourhood_cleansed', 'neighbourhood_group_cleansed', 'latitude', 'longitude', 'room_type', 'bedrooms', 'beds', 'number_of_reviews', 'availability_365','price']]
export_df.to_csv(saved_file, index = False, header=True)

print(export_df)
