==> Price Insight System/4.final linear model.ipynb <==
# Basic Libraries
import numpy as np
import pandas as pd
import seaborn as sb
import matplotlib.pyplot as plt # we only need pyplot
sb.set() # set the default Seaborn style for graphics

from wordcloud import WordCloud
#if encounter error: No module named 'wordcloud'
#run the line below

import folium
from folium.plugins import HeatMap
#if encounter error: No module named 'folium'
#run the line below

import plotly.offline as pyo
import plotly.graph_objs as go
#if encounter error: No module named 'plotly'
#run the line below

import seaborn as sns
sns.set_style('whitegrid')

# Import the dataset
airbnb_data = pd.read_csv("PredictPrice.csv")
airbnb_df = pd.DataFrame(airbnb_data)
airbnb_data.head(5)

airbnb_data[['price','new_host_acceptance_rate', 'bedrooms', 'beds', 'number_of_reviews', 'availability_365']].describe()

__we can see a huge difference in the range of values present in our numerical features:
price,beds,number of reviews and availability_365. We will use scaling to fix this problem__
__to be done during the model building phase__

things to do be done before model building
1) Normalization is a scaling technique in which values are shifted and rescaled so that they end up ranging between 0 and 1. It is also known as Min-Max scaling. Feature scaling to be done later when tuning model for comparsion
2) Convert categorical variable into numbers some  ML libraries do not take categorical variables as input. Label encoder (neigbourhood and region)

airbnb_df

#remove rows that are zero
airbnb_df = airbnb_df.loc[(airbnb_df['price'] > 0)]

from sklearn import preprocessing
category_col =['neighbourhood_cleansed', 'neighbourhood_group_cleansed', 'room_type','host_is_superhost']
labelEncoder = preprocessing.LabelEncoder()
mapping_dict ={}
for col in category_col:
    airbnb_df[col] = labelEncoder.fit_transform(airbnb_df[col])
    le_name_mapping = dict(zip(labelEncoder.classes_,
                        labelEncoder.transform(labelEncoder.classes_)))
    mapping_dict[col]= le_name_mapping
print(mapping_dict)


airbnb_df

new_df=airbnb_df

##log transfromation
new_df.availability_365 += 0.000000001
new_df['availability_365'] = np.log10(new_df['availability_365'])

new_df.number_of_reviews += 0.000000001
new_df['number_of_reviews'] = np.log10(new_df['number_of_reviews'])

indexidtracker = new_df[['id']]
latitudetracker = new_df[['latitude']]
longitudetracker = new_df[['longitude']]
##model 1 all variable
Model1  = new_df[['new_host_acceptance_rate','host_is_superhost','room_type','bedrooms'\
                  ,'beds','number_of_reviews','availability_365','neighbourhood_cleansed','neighbourhood_group_cleansed']]
##Model 2 room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat
Model2  = new_df[['room_type','bedrooms'\
                  ,'number_of_reviews','neighbourhood_cleansed','neighbourhood_group_cleansed']]
##model 3 room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat
Model3  = new_df[['room_type','number_of_reviews','neighbourhood_group_cleansed']]
y = new_df['price'].values
y = np.log10(y)


# Import train_test_split from sklearn
from sklearn.model_selection import train_test_split
m1_train, m1_test,m2_train,m2_test,m3_train,m3_test,y_train,y_test,id_train,id_test,lat_train,lat_test,long_train,long_test = train_test_split(Model1,Model2,Model3,y,indexidtracker,latitudetracker,longitudetracker,test_size = 0.2)

# Check the sample sizes for salesprice
print("Train Set :", m1_train.shape, y_train.shape)
print("Test Set  :", m1_test.shape, y_test.shape)
# Check the sample sizes overallqual
print("Train Set :", m2_train.shape, y_train.shape)
print("Test Set  :", m2_test.shape, y_test.shape)
# Check the sample sizes year built
print("Train Set :", m3_train.shape, y_train.shape)
print("Test Set  :", m3_test.shape, y_test.shape)

# Fitting Multiple Linear Regression to the Training set
from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(m1_train, y_train)
# Predicting the Test set results
y_pred = lr.predict(m1_test)

model1df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_pred, 0)})
model1df.head(10)

from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import r2_score

print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m1_test))), 2))
m1rmse = np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m1_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, lr.predict(m1_train), multioutput='variance_weighted'), 2))
m1r2train= np.round(r2_score(y_train, lr.predict(m1_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, lr.predict(m1_test), multioutput='variance_weighted'), 2))
m1r2test= np.round(r2_score(y_test, lr.predict(m1_test), multioutput='variance_weighted'), 2)

## Model 2 room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat

# Fitting Multiple Linear Regression to the Training set
from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(m2_train, y_train)
# Predicting the Test set results
y_pred2 = lr.predict(m2_test)

model2df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_pred2, 0)})
model2df.head(10)

print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m2_test))), 2))
m2rmse = np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m2_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, lr.predict(m2_train), multioutput='variance_weighted'), 2))
m2r2train= np.round(r2_score(y_train, lr.predict(m2_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, lr.predict(m2_test), multioutput='variance_weighted'), 2))
m2r2test= np.round(r2_score(y_test, lr.predict(m2_test), multioutput='variance_weighted'), 2)


## Model 3 room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat

# Fitting Multiple Linear Regression to the Training set
from sklearn.linear_model import LinearRegression
lr = LinearRegression()
lr.fit(m3_train, y_train)
# Predicting the Test set results
y_pred3 = lr.predict(m3_test)

model3df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_pred3, 0)})
model3df.head(10)

print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m3_test))), 2))
m3rmse = np.round(np.sqrt(metrics.mean_squared_error(y_test, lr.predict(m3_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, lr.predict(m3_train), multioutput='variance_weighted'), 2))
m3r2train =  np.round(r2_score(y_train, lr.predict(m3_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, lr.predict(m3_test), multioutput='variance_weighted'), 2))
m3r2test= np.round(r2_score(y_test, lr.predict(m3_test), multioutput='variance_weighted'), 2)

## Using Random forest to optimize

## Model 1 all variable

# Fitting Random Forest Regression to the dataset
from sklearn.ensemble import RandomForestRegressor
rfr = RandomForestRegressor(max_depth=8, n_estimators = 100, random_state = 0)
rfr.fit(m1_train, y_train)
# Predicting the Test set results
y_predr1 = rfr.predict(m1_test)
df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_predr1, 0)})
df.head(10)

from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import r2_score
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m1_test))), 2))
m1rmserr =  np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m1_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, rfr.predict(m1_train), multioutput='variance_weighted'), 2))
m1r2trainrr=  np.round(r2_score(y_train, rfr.predict(m1_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, rfr.predict(m1_test), multioutput='variance_weighted'), 2))
m1r2testrr = np.round(r2_score(y_test, rfr.predict(m1_test), multioutput='variance_weighted'), 2)

## Model 2 room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat

# Fitting Random Forest Regression to the dataset
from sklearn.ensemble import RandomForestRegressor
rfr = RandomForestRegressor(max_depth=8, n_estimators = 100, random_state = 0)
rfr.fit(m2_train, y_train)
# Predicting the Test set results
y_predr2 = rfr.predict(m2_test)
df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_predr2, 0)})
df.head(10)

from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import r2_score
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m2_test))), 2))
m2rsmerr= np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m2_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, rfr.predict(m2_train), multioutput='variance_weighted'), 2))
m2r2trainrr= np.round(r2_score(y_train, rfr.predict(m2_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, rfr.predict(m2_test), multioutput='variance_weighted'), 2))
m2r2testrr=np.round(r2_score(y_test, rfr.predict(m2_test), multioutput='variance_weighted'), 2)

## Model 3 room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat

# Fitting Random Forest Regression to the dataset
from sklearn.ensemble import RandomForestRegressor
rfr = RandomForestRegressor(max_depth=8, n_estimators = 100, random_state = 0)
rfr.fit(m3_train, y_train)
# Predicting the Test set results
y_predr3 = rfr.predict(m3_test)
df = pd.DataFrame({'Actual': np.round(10 ** y_test, 0), 
                   'Predicted': np.round(10 ** y_predr3, 0)})
df.head(10)

from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import r2_score
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:', np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m3_test))), 2))
m3rsmerr=np.round(np.sqrt(metrics.mean_squared_error(y_test, rfr.predict(m3_test))), 2)
print('R2 score train:', np.round(r2_score(y_train, rfr.predict(m3_train), multioutput='variance_weighted'), 2))
m3r2trainrr = np.round(r2_score(y_train, rfr.predict(m3_train), multioutput='variance_weighted'), 2)
print('R2 score test:', np.round(r2_score(y_test, rfr.predict(m3_test), multioutput='variance_weighted'), 2))
m3r2testrr =  np.round(r2_score(y_test, rfr.predict(m3_test), multioutput='variance_weighted'), 2)

## Selecting the Best Model For Data Driven Insight and Demonstration of concept using flask

from sklearn.metrics import mean_squared_error
from sklearn import metrics
from sklearn.metrics import r2_score
#Comparing Models
print("Model 1 All variable")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m1rmse)
print('R2 score train:',m1r2train)
print('R2 score test:', m1r2test)
#RandomForest model 1 (All variable)
print("Model 1 All RR")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m1rmserr)
print('R2 score train:',m1r2trainrr)
print('R2 score test:', m1r2testrr)
#Multivariate regression model 2(room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat)
print("Model 2 room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m2rmse)
print('R2 score train:',m2r2train)
print('R2 score test:', m2r2test)
#RandomForest model 2 (room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat)
print("Model 2 RR room_type_cat, bedroom,number_of_reviews,neighbourhood_cleansed_cat, neighbourhood_group_cleansed_cat")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m2rsmerr)
print('R2 score train:',m2r2trainrr)
print('R2 score test:', m2r2testrr)
#Multivariate regression model 3 (room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat)
print("Model 3 room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m3rmse)
print('R2 score train:',m3r2train)
print('R2 score test:', m3r2test)
#randomforest model 3  (room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat)
print("Model 3 RR room_type_cat,number_of_reviews,neighbourhood_group_cleansed_Cat")
print("********************************************************************************************************************")
print('Price mean:', np.round(np.mean(y), 2))  
print('Price std:', np.round(np.std(y), 2))
print('RMSE:',m3rsmerr)
print('R2 score train:',m3r2trainrr)
print('R2 score test:', m3r2testrr)


__Conclusion__
1) Price prediction are performing well after using random forest model compared to multivariate regresion. Best score is 0.57(RR) vs 0.37(MVR)
2) Using categorical encoded data did improve the model.
3) Using more variable increase the accuracy of the model showing that, all the variable plays an important role in predicting price for property as there are a lot of factors affecting property pricing.
4) Scaling value and logging values improve model performance because , omitting scaling will basically result in models that are disproportionally influenced by the subset of features on a large scale.
5) When there are large number of features with less data-sets(with low noise), linear regressions may outperform Decision trees/random forests. But for this project random forest perform better because decision tree have better average accuracy. The model also used alot of categorical variable thus random forest perfom bettr.

__Random Forest Model 1 is the best model because the highest r2 and lowest rmse.__ Used by Airbnb
RMSE: 0.23
R2 score train: 0.7
R2 score test: 0.57
R2 score of 0.57 suggest that 57% represents a model that explains all the variation in the response variable around its mean. 
A low RMSE value of 0.23 indicates that the simulated and observed data are close to each other showing a better accuracy

__Random Forest Model 2 is the second best model__ Used for developing website for Airbnb host

__Deployment is one of the last stages of any machine learning project. This model have two different end-users, Airbnb host and Airbnb company
1) For Airbnb host, the team decided to target new Airbnb host. This model aims to help the new Airbnb host to guage/price their property at reasonable market rate. The model will be deployed using flask. A website will be provided for the user to input their proprety features and the website will output the predicted price. The team will deploy the second best model to this application because the team feel that that the best model should be used by Airbnb company as this model will give them competitive advantage over their competitors in the industry.

2) For Airbnb company, the model aims to help the company forecast revenue and guage whether the current price stragety is reasonable assuming the model is correct.

##Saving model into pickle for Airbnb host flask application

##Model to be used for flask
import pickle
rfr = RandomForestRegressor(max_depth=8, n_estimators = 100, random_state = 0)
rfr.fit(m2_train, y_train)
#make pickle file for our model
pickle.dump(rfr,open("model.pkl", 'wb'))

## generating predicted dataframe using the best model

rfr = RandomForestRegressor(max_depth=8, n_estimators = 100, random_state = 0)
rfr.fit(m1_train, y_train)
y_predr1 = rfr.predict(m1_test)
df = pd.DataFrame({'Actual Price': np.round(10 ** y_test, 0), 
                  'Predicted Price': np.round(10 ** y_predr1, 0)})


#to be concat later
df

m2_test

##logging back all the different variable
exportm1test = m1_test
exportm1test.number_of_reviews = np.round(10 ** exportm1test.number_of_reviews, 0)
exportm1test.availability_365 = np.round(10 ** exportm1test.availability_365, 0)

id_test

## inverting dictionary back to use dictionary to convert numerical to categorical

map1 = mapping_dict['neighbourhood_cleansed']
inv_map1 = {v: k for k, v in map1.items()}
exportm1test['neighbourhood_cleansed'].replace(inv_map1, inplace=True)

map2 = mapping_dict['neighbourhood_group_cleansed']
inv_map2 = {v: k for k, v in map2.items()}
exportm1test['neighbourhood_group_cleansed'].replace(inv_map2, inplace=True)

map3 = mapping_dict['room_type']
inv_map3 = {v: k for k, v in map3.items()}
exportm1test['room_type'].replace(inv_map3, inplace=True)

map4 = mapping_dict['host_is_superhost']
inv_map4 = {v: k for k, v in map4.items()}
exportm1test['host_is_superhost'].replace(inv_map4, inplace=True)

exportm1test['id']=id_test['id']

df.reset_index(drop=True, inplace=True)
exportm1test.reset_index(drop=True, inplace=True)

finaldataframe = (pd.concat([df, exportm1test], axis=1))
tempdataframe = (pd.concat([lat_test,long_test],axis=1))

finaldataframe.reset_index(drop=True, inplace=True)
tempdataframe.reset_index(drop=True, inplace=True)

bigframe = (pd.concat([finaldataframe,tempdataframe], axis=1))

bigframe

bigframe.to_csv('model1predicted.csv')
