==> Price Insight System/5.Data_Driven_Insight.ipynb <==
# Basic Libraries
import pandas as pd
import matplotlib.pyplot as plt
import seaborn as sns
from wordcloud import WordCloud
import re

#importing
predicted_frame= pd.read_csv('model1predicted.csv')

predicted_frame

lessthan50frame =  predicted_frame[(predicted_frame['Actual Price'] <50)] 

lessthan50frame

%%html
<iframe src="https://e.infogram.com/e76e0dea-13a1-45df-8a49-d0cec207ab2e?src=embed" title="PGSG_2021_June_Article-EN_49922_Condo Rental Prices by units" width="700" height="972" scrolling="no" frameborder="0" style="border:none;" allowfullscreen="allowfullscreen"></iframe>

__Assumptions__
Several assumptions was made to derive this datadriven insight. The team assume that the best random forest model selected provide the recommended price compared to the host. 
The justification why the model provided a better price is shown below
1) observe the fifth row for lessthan50frame. This property is a private 1 room condo, 1 bedroom located in geylang central region. geylang belongs to district 19. Refering to propertyguru property website, the average price of such room is 1619 per month which is 52 dollar per day. The price set by user is 39 dollar while the predicted price is 54 dollar. This shows that the model recommend price that is close to the market rate. Simillar trend can be found in row 4, located in pasir ris east region district 18 which is 1442 per month which is 80 dollar per day, the predicted price is 91 while the price set by user is 30.	
__Thus the team will based on our Data_Driven_insight assuming the random forest model predict a better value compared to the host(actual price). We will compare the actual price and predicted price and generate meaningful insights. 

##Forecasting venue per month using availability_365
##availability_365 is how many day the room is rent,we will derive forecast venue per month by usin this formula
#divide by 12 is 12 month
#forecast = (availbility_365*price)/12  
sum_column_actual = round((predicted_frame["availability_365"] * predicted_frame["Actual Price"])/12,2)
predicted_frame["forecast_per_month_actual"] = sum_column_actual 

sum_column_predicted = round((predicted_frame["availability_365"] * predicted_frame["Predicted Price"])/12,2)
predicted_frame["forecast_per_month_Predicted Price"] = sum_column_predicted 
actualvspredict = round((predicted_frame["forecast_per_month_Predicted Price"] - predicted_frame["forecast_per_month_actual"]),2)
predicted_frame["difference"] = actualvspredict

__forecast per month predicted price and forecast per month actual price__
To forecast the monthly revenue of each property, the team decided to use availability_365, to guage how many days the room is rent and this value will be multiply by the price divided by 12.
The term actual price used is refer to the price set by the host.
the term predicted price is the recommended price by the model.
The variable difference is deviered from predicted price- actual price to see the difference

predicted_frame

predicted_frame.to_csv('model1predicted_final.csv')
